requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.25.0
PyPDF2>=3.0.0
python-multipart>=0.0.5
pydantic>=2.0.0
openai>=1.0.0
sentence-transformers>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0